+++ lib/pro_image_editor.dart
@@ -23,9 +23,9 @@ export 'package:pro_image_editor/models/import_export/import_state_history.dart'
 export 'package:pro_image_editor/models/import_export/import_state_history_configs.dart';
 export 'package:pro_image_editor/models/import_export/utils/export_import_enum.dart';
 
+export 'package:pro_image_editor/models/aspect_ratio_item.dart';
 export 'package:pro_image_editor/utils/design_mode.dart';
 export 'package:pro_image_editor/modules/paint_editor/utils/paint_editor_enum.dart';
-export 'package:pro_image_editor/modules/crop_rotate_editor/utils/crop_aspect_ratios.dart';
 export 'package:pro_image_editor/widgets/layer_widget.dart'
     show LayerBackgroundColorModeE;